JS15/daynight-classifier-IrsyadDimas/requirements.txt
flask
tensorflow-cpu
numpy
scikit-learn
scikit-image
opencv-python-headless
gunicorn